==> Rmd/01-aggregate-genos-and-gsi-to-species.Rmd <==
---
title: "Aggregrate Multiply-genotyped indivs and assign kelp rockfish with GSI"
output: 
  html_notebook:
    toc: true
---




To start off with, let's load data and libs:
```{r load-stuff}
library(tidyverse)
library(CKMRsim)
library(stringr)

meta <- readRDS("../data/processed/meta-data-tibble.rds")
genos <- readRDS("../extdata/processed/called_genos_na_explicit.rds") %>%
  filter(NMFS_DNA_ID %in% meta$NMFS_DNA_ID)  # drop those we don't have meta data for
samples <- readRDS("../data/processed/sample-sheet-tibble.rds") %>%
  filter(NMFS_DNA_ID %in% meta$NMFS_DNA_ID)
```


## Some initial filters

### Only certain species complexes

Now, get only adults and juveniles of certain species or complexes.  Basically toss the reported _melanops_ indivs.
```{r get-spp}
keepers <- meta %>%
  filter(REPORTED_LIFE_STAGE %in% c("ADULT", "JUVENILE") &
           !(SPECIES=="melanops")) %>%
  select(NMFS_DNA_ID) %>%
  unlist() %>% unname()
```

Whittle the genotype data down to just those:
```{r whittle-genos}
kgenos <- genos %>%
  filter(NMFS_DNA_ID %in% keepers)
```

### Take highest read-depth call for multiply-genotyped DNA_IDs

Now, here is a harder operation: if an individual is multiply-genotyped, take the
genotype with the highest total read depth.  
```{r take-just-one}
# slow-ish function to get the total read depth column
tdepth <- function(a, d) {
  if(any(is.na(a))) {
    return(NA)
  }
  if(a[1]==a[2]) {
    return(d[1])
  } else {
    return(d[1] + d[2])
  }
  
}
# this takes the highest read-depth instance of each duplicately-genotyped individual.
geno_one_each <- kgenos %>%
  group_by(NMFS_DNA_ID, locus, gtseq_run, id) %>%
  mutate(total_depth = tdepth(allele, depth)) %>%
  ungroup() %>%
  arrange(NMFS_DNA_ID, locus, total_depth, gtseq_run, id, depth) %>%
  group_by(NMFS_DNA_ID, locus) %>%
  mutate(rank = 1:n()) %>%
  ungroup() %>%
  filter(rank <= 2)
```

### Toss out indivs with fewer than 85 non-missing loci
Now, toss out any individual with fewer than 85 non-missing loci
```{r toss-missers}
no_hi_missers <- geno_one_each %>% 
  group_by(NMFS_DNA_ID) %>%
  filter(sum(!is.na(allele)) >= (85*2))
```
So, we started with `r length(unique(geno_one_each$NMFS_DNA_ID))` from our target species
and after filtering out indivs with fewer than 85 genotyped loci, we were left with 
`r length(unique(no_hi_missers$NMFS_DNA_ID))` individuals.  Those are the ones that
we will run through gsi_sim to identify to species.



## Making a GSI data set

We are going to do this by turning alleles into integers and spreading it and then using 
gsi_sim.
```{r spread-genos}
# first make integers of the alleles
alle_idxs <- no_hi_missers %>% 
  select(NMFS_DNA_ID, locus, gene_copy, allele) %>%
  group_by(locus) %>%
  mutate(alleidx = as.integer(factor(allele, levels = unique(allele)))) %>%
  ungroup() %>%
  arrange(NMFS_DNA_ID, locus, alleidx)

# then make a space-separated string of the genotype
genostrs <- alle_idxs %>% 
  group_by(NMFS_DNA_ID, locus) %>%
  summarise(genostr = paste(alleidx[1], alleidx[2], collapse = " ")) %>%
  mutate(genostr = ifelse(genostr == "NA NA", "0 0", genostr))  # denote missing alleles with 0
  
# finally spread that and we have a data frame that is
# close to final
gsi_frame <- genostrs %>%
  tidyr::spread(data = ., key = locus, value = genostr)
  
```


Once we have that, we just need to assign integers to figure out whom we are
going to use.  For the baselines, I say we make a data set of all the species that
have adults with appreciable numbers of individuals.

```{r make-group-ints}
meta_keep <- meta %>%
  filter(NMFS_DNA_ID %in% genostrs$NMFS_DNA_ID) %>%
  group_by(REPORTED_LIFE_STAGE, SPECIES) 

atro_groups <- meta_keep %>%
  tally() %>%
  filter(n>=20,
         REPORTED_LIFE_STAGE == "ADULT" | str_detect(SPECIES, "atrovirens"))

# see what that looks like
atro_groups
```
That is `r sum(atro_groups$n)` that we will be working with.  Some of them will be kelp rockfish.  Cool.

Let's pick them all out:
```{r pick-out}
atro_dults <- atro_groups %>% filter(REPORTED_LIFE_STAGE == "ADULT")
atro_juvs <- atro_groups %>% filter(REPORTED_LIFE_STAGE == "JUVENILE")

adult_ids <- meta_keep %>%
  ungroup() %>%
  filter(REPORTED_LIFE_STAGE == "ADULT" & SPECIES %in% atro_dults$SPECIES) %>%
  mutate(group = SPECIES) %>%
  select(NMFS_DNA_ID, group) %>%
  arrange(group)

juvie_ids <- meta_keep %>%
  ungroup() %>%
  filter(REPORTED_LIFE_STAGE == "JUVENILE" & SPECIES %in% atro_juvs$SPECIES) %>%
  mutate(group = paste(REPORTED_LIFE_STAGE, SPECIES, sep = "_")) %>%
  select(NMFS_DNA_ID, group) %>%
  arrange(group)

# now, weird thing here is that this leads to duplicated IDs for the juveniles, 
# and could also do so for the adults, but we
# only want to pick it out once, so we have to condense them here
juvie_ids <- juvie_ids %>%
  group_by(NMFS_DNA_ID) %>%
  summarise(group = "JUVENILE_atrovirens")
```

Now we just need to grab those individauls out of the gsi_frame and make some gsi_sim files:
```{r make-adults-gsi}
adult_genos <- left_join(adult_ids, gsi_frame)
fishlist <- split(adult_genos %>% select(-group), adult_genos$group)
cat(c(nrow(adult_genos), ncol(adult_genos) - 2), sep = " ", eol = "\n", file = "../gsi_area/adults.txt")
cat(names(adult_genos[-(1:2)]), sep = "\n", file = "../gsi_area/adults.txt", append = TRUE)
lapply(names(fishlist), function(n) {
  y <- fishlist[[n]]
  cat(c("POP", n), sep = " ", eol = "\n", file = "../gsi_area/adults.txt", append = TRUE)
  write.table(y, row.names = F, col.names = F, quote = F, sep = "   ", file = "../gsi_area/adults.txt", append = TRUE)
})
```

And I do something similar for the juvies
```{r make-juvies-gsi}
juvie_genos <- left_join(juvie_ids, gsi_frame)
fishlist <- split(juvie_genos %>% select(-group), juvie_genos$group)
cat(c(nrow(juvie_genos), ncol(juvie_genos) - 2), sep = " ", eol = "\n", file = "../gsi_area/juvies.txt")
cat(names(adult_genos[-(1:2)]), sep = "\n", file = "../gsi_area/juvies.txt", append = TRUE)
cat(c("POP Mixture"), eol = "\n", file = "../gsi_area/juvies.txt", append = TRUE)
lapply(names(fishlist), function(n) {
  y <- fishlist[[n]]
  write.table(y, row.names = F, col.names = F, quote = F, sep = "   ", file = "../gsi_area/juvies.txt", append = TRUE)
})
```

Then run gsi_sim on the command line.  This is set to not eval by default.  
```{sh run-gsi-sim, eval=FALSE}
# first we do the juvenile assignments
cd gsi_area; 
../bin/gsi_sim-Darwin -b adults.txt -t juvies.txt | awk -F";" 'BEGIN {print "NMFS_DNA_ID species score"} /GMA_FULL_EM_INDIVS_CSV:/ {print $1, $2, $3}' | sed 's/GMA_FULL_EM_INDIVS_CSV:\///g' > juvenile_rockfish_assignments.txt

# then we do self-assignment to weed out any adults that are misidentified to species
../bin/gsi_sim-Darwin -b adults.txt --self-assign | awk -F";" 'BEGIN {print "NMFS_DNA_ID species score"} /SELF_ASSIGN_A_LA_GC_CSV:/ {print $1, $2, $3}' | sed 's/SELF_ASSIGN_A_LA_GC_CSV:\///g;' > adult_rockfish_self_assignment.txt 

```

```{r slurp-gsi, eval=FALSE}
juvie_rock_ass <- read_delim("../gsi_area/juvenile_rockfish_assignments.txt", delim = " ")
saveRDS(juvie_rock_ass, file = "../data/processed/juvie-gsi-sim-assignments.rds", compress = "xz")

adult_rock_ass <- read_delim("../gsi_area/adult_rockfish_self_assignment.txt", delim = " ")
saveRDS(adult_rock_ass, file = "../data/processed/adult-gsi-sim-self-assignments.rds", compress = "xz")
```

That has given us a nice little file that gives us the gsi-sim assignments, which should be good enough
for us to get the kelp rockfish.  

As a final hurrah, we will also save no-high-missers
```{r save-no-hi}
saveRDS(no_hi_missers, file = "../extdata/processed/genos-aggregated-and-no-hi-missers.rds", compress = "xz")
```
